+++ HISTORY.md
@@ -2,6 +2,14 @@
 Release History
 ===============
 
+0.3 (2020-10-29)
+----------------
+
+- Add Python 3.9
+- Rework exception management to introduce an error code and remove one exception class
+- Add option to sort tags by numeric order
+
+
 0.2 (2020-03-22)
 ----------------